fdpgen initPositions: squash C4244 warning
MSVC notes:

  graphviz\lib\fdpgen\tlayout.c(522,15): warning C4244: '=': conversion from
    'time_t' to 'long', possible loss of data

This code is just getting a value to seed a random number generator, so
truncation is fine.

+++ lib/fdpgen/tlayout.c
@@ -519,7 +519,7 @@ static pointf initPositions(graph_t * g, bport_t * pp)
 	local_seed = T_seed;
     else {
 #if defined(_WIN32)
-	local_seed = time(NULL);
+	local_seed = (long)time(NULL);
 #else
 	local_seed = getpid() ^ time(NULL);
 #endif
